requirements.txt
fastapi
uvicorn
pydantic
httpx
httpx
python-dotenv
supabase
pytest
google-generativeai
google-genai
agno
